requirements.txt
langgraph
pydantic
pandas
duckdb
hdbcli
google-cloud-bigquery
langchain-google-genai
python-dotenv